+++ containers/Dashboard/locales/en.json
@@ -272,10 +272,10 @@
   "dashborad.server_project_numbers": "The distribution of the number of hosts for each project",
   "dashboard.group_by": "Group By",
   "dashboard.optimization_suggestion": "Optimization suggestions",
-  "dashboard.optimization_suggestion_host_up": "宿主机可升配",
-  "dashboard.optimization_suggestion_host_down": "宿主机可降配",
-  "dashboard.optimization_suggestion_cloudhost_up": "云主机可升配",
-  "dashboard.optimization_suggestion_cloudhost_down": "云主机可降配",
+  "dashboard.optimization_suggestion_host_up": "Uploadable host",
+  "dashboard.optimization_suggestion_host_down": "Downloadable host",
+  "dashboard.optimization_suggestion_cloudhost_up": "Uploadable Virtual Machine",
+  "dashboard.optimization_suggestion_cloudhost_down": "Downloadable Virtual Machine",
   "dashboard.text_191": "CPU",
   "dashboard.text_192": "Memory",
   "dashboard.monitor": "Monitoring information",
