+++ containers/Dashboard/styles/index.less
@@ -6,7 +6,6 @@
   height: 100%;
   background-color: #fff;
   padding: 15px;
-  border-radius: 5px;
 }
 
 .dashboard-card-header {
@@ -14,10 +13,8 @@
   flex-shrink: 0;
   flex-grow: 0;
   align-items: center;
-  font-size: 16px;
-  font-family: PingFangSC-Medium, PingFang SC;
-  font-weight: 500;
-  color: rgba(0, 0, 0, 0.85);
+  margin-bottom: 10px;
+  color: @gray_text_1;
 }
 
 .dashboard-card-header .dashboard-card-header-left {
