+++ containers/Dashboard/styles/index.scss
@@ -5,7 +5,6 @@
   height: 100%;
   background-color: #fff;
   padding: 15px;
-  border-radius: 5px;
 }
 
 .dashboard-card-header {
@@ -13,10 +12,7 @@
   flex-shrink: 0;
   flex-grow: 0;
   align-items: center;
-  font-size: 16px;
-  font-family: PingFangSC-Medium, PingFang SC;
-  font-weight: 500;
-  color: rgba(0, 0, 0, 0.85);
+  margin-bottom: 10px;
 }
 
 .dashboard-card-header .dashboard-card-header-left {
